Refactor internals to split out modules
