+++ .isort.cfg
@@ -1,7 +1,7 @@
 [settings]
 default_section=THIRDPARTY
 indent='    '
-sections=FUTURE,STDLIB,THIRDPARTY,FIRSTPARTY,TESTS,LOCALFOLDER
+sections=FUTURE,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER
 multi_line_output=3
 line_length=88
 include_trailing_comma=True
