+++ onfido/models/__init__.py
@@ -1,0 +1,6 @@
+from .applicant import Applicant
+from .check import Check
+from .event import Event
+from .report import Report
+
+__all__ = ["Applicant", "Check", "Report", "Event"]
